extractChoose #8 - Drop PlayerChooser for DataChooserBuilder
